Moved input parts to internal directory
